requirements.txt
selenium
python-dotenv
langchain
langchain-google-genai